requirements.txt
# core
Django<1.12,>=1.11

# static files
django_compressor<2.2,>=2.1
django-libsass<0.8,>=0.7

# api
djangorestframework==3.6.2

# testing
# should have flake8 and coveralls too but will split into another file
# faker is reqired by a core management command
Faker<0.8,>=0.7
